Add scan mode function separate to main loop
