+++ firmware/hackrf_usb/hackrf_usb.c
@@ -22,8 +22,6 @@
 
 #include <stddef.h>
 
-#include <libopencm3/cm3/vector.h>
-
 #include <libopencm3/lpc43xx/m4/nvic.h>
 
 #include <streaming.h>
@@ -45,73 +43,8 @@
 #include "usb_api_scan.h"
 
 #include "usb_api_transceiver.h"
-#include "sgpio_isr.h"
 #include "usb_bulk_buffer.h"
  
-static volatile transceiver_mode_t _transceiver_mode = TRANSCEIVER_MODE_OFF;
-
-void set_transceiver_mode(const transceiver_mode_t new_transceiver_mode) {
-	baseband_streaming_disable(&sgpio_config);
-	
-	usb_endpoint_disable(&usb_endpoint_bulk_in);
-	usb_endpoint_disable(&usb_endpoint_bulk_out);
-	
-	_transceiver_mode = new_transceiver_mode;
-	
-	if( _transceiver_mode == TRANSCEIVER_MODE_RX ) {
-		led_off(LED3);
-		led_on(LED2);
-		usb_endpoint_init(&usb_endpoint_bulk_in);
-		rf_path_set_direction(&rf_path, RF_PATH_DIRECTION_RX);
-		vector_table.irq[NVIC_SGPIO_IRQ] = sgpio_isr_rx;
-	} else if (_transceiver_mode == TRANSCEIVER_MODE_TX) {
-		led_off(LED2);
-		led_on(LED3);
-		usb_endpoint_init(&usb_endpoint_bulk_out);
-		rf_path_set_direction(&rf_path, RF_PATH_DIRECTION_TX);
-		vector_table.irq[NVIC_SGPIO_IRQ] = sgpio_isr_tx;
-	} else {
-		led_off(LED2);
-		led_off(LED3);
-		rf_path_set_direction(&rf_path, RF_PATH_DIRECTION_OFF);
-		vector_table.irq[NVIC_SGPIO_IRQ] = sgpio_isr_rx;
-	}
-
-	if( _transceiver_mode != TRANSCEIVER_MODE_OFF ) {
-		si5351c_activate_best_clock_source(&clock_gen);
-		baseband_streaming_enable(&sgpio_config);
-	}
-}
-
-transceiver_mode_t transceiver_mode(void) {
-	return _transceiver_mode;
-}
-
-usb_request_status_t usb_vendor_request_set_transceiver_mode(
-	usb_endpoint_t* const endpoint,
-	const usb_transfer_stage_t stage
-) {
-	if( stage == USB_TRANSFER_STAGE_SETUP ) {
-		switch( endpoint->setup.value ) {
-		case TRANSCEIVER_MODE_OFF:
-		case TRANSCEIVER_MODE_RX:
-		case TRANSCEIVER_MODE_TX:
-			set_transceiver_mode(endpoint->setup.value);
-			usb_transfer_schedule_ack(endpoint->in);
-			return USB_REQUEST_STATUS_OK;
-		case TRANSCEIVER_MODE_CPLD_UPDATE:
-			usb_endpoint_init(&usb_endpoint_bulk_out);
-			start_cpld_update = true;
-			usb_transfer_schedule_ack(endpoint->in);
-			return USB_REQUEST_STATUS_OK;
-		default:
-			return USB_REQUEST_STATUS_STALL;
-		}
-	} else {
-		return USB_REQUEST_STATUS_OK;
-	}
-}
-
 static const usb_request_handler_fn vendor_request_handler[] = {
 	NULL,
 	usb_vendor_request_set_transceiver_mode,
@@ -247,50 +180,44 @@ int main(void) {
 	rf_path_init(&rf_path);
 
 	unsigned int phase = 0;
-	unsigned int blocks_queued = 0;
 
 	while(true) {
 		// Check whether we need to initiate a CPLD update
 		if (start_cpld_update)
 			cpld_update();
 
+		// Check whether we need to initiate scan mode
+		if (start_scan_mode)
+			scan_mode();
+
 		// Set up IN transfer of buffer 0.
 		if ( usb_bulk_buffer_offset >= 16384
 		     && phase == 1
 		     && transceiver_mode() != TRANSCEIVER_MODE_OFF) {
-			if (blocks_queued == 2)
-				usb_transfer_schedule_block(
-					(transceiver_mode() == TRANSCEIVER_MODE_RX)
-					? &usb_endpoint_bulk_in : &usb_endpoint_bulk_out,
-					&usb_bulk_buffer[0x0000],
-					0x4000,
-					NULL, NULL
-					);
+			usb_transfer_schedule_block(
+				(transceiver_mode() == TRANSCEIVER_MODE_RX)
+				? &usb_endpoint_bulk_in : &usb_endpoint_bulk_out,
+				&usb_bulk_buffer[0x0000],
+				0x4000,
+				NULL, NULL
+				);
 			phase = 0;
-			blocks_queued++;
 		}
-	
+
 		// Set up IN transfer of buffer 1.
 		if ( usb_bulk_buffer_offset < 16384
 		     && phase == 0
 		     && transceiver_mode() != TRANSCEIVER_MODE_OFF) {
-			if (blocks_queued == 2)
-				usb_transfer_schedule_block(
-					(transceiver_mode() == TRANSCEIVER_MODE_RX)
-					? &usb_endpoint_bulk_in : &usb_endpoint_bulk_out,
-					&usb_bulk_buffer[0x4000],
-					0x4000,
-					NULL, NULL
-				);
+			usb_transfer_schedule_block(
+				(transceiver_mode() == TRANSCEIVER_MODE_RX)
+				? &usb_endpoint_bulk_in : &usb_endpoint_bulk_out,
+				&usb_bulk_buffer[0x4000],
+				0x4000,
+				NULL, NULL
+			);
 			phase = 1;
-			blocks_queued++;
-		}
-
-		if (blocks_queued > 2) {
-			scan_callback();
-			blocks_queued = 0;
 		}
 	}
-	
+
 	return 0;
 }
